# refactor(rslint_parser): Use `f64` and `i64` to parse numbers, remove lexical