--- a/crates/rslint_parser/Cargo.toml
+++ b/crates/rslint_parser/Cargo.toml
@@ -13,7 +13,6 @@ rslint_syntax = { path = "../rslint_syntax", version = "0.1" }
 rslint_lexer = { path = "../rslint_lexer", version = "0.2", features = ["highlight"] }
 rome_rowan = { path = "../rome_rowan", version = "0.0.0" }
 num-bigint = "0.4.3"
-lexical = { version = "5.2.0", features = ["radix"] }
 drop_bomb = "0.1.5"
 bitflags = "1.3.2"
 indexmap = "1.8.0"