## Summary
Number parsing is incomplete at the moment. This is the first part of on going effort to improve number parsing.

Previously rslint used the `lexical` library to parse any number into f64 without checking, and has a double parsing issue when parsing numbers with leading zeroes.

This PR only removes `lexical`, and uses
* `f64` to parse numbers with radix 10
* `i64` to parse other radix 2, 8 and 16, since JavaScript does not have floating point binary/octal/hex numbers. 
 
Correctness checking and other improvements will be submitted in other PRs.

## Test Plan
All tests pass.