@MichaReiser 2022-01-19
Would you mind adding a few examples of what this PR fixes? I'm having troubles understanding why the change is needed, especially because there are no new tests that were failing before but are fixed now.