@Boshen 2022-01-19
This is just a removable of an unnecessary library. No behavior changed should be introduced in this PR.

Rslint needed lexical because float parsing was slow and incomplete in rust, but everything is now fixed under `https://github.com/rust-lang/rust/pull/86761`. The author of lexical moved all standard floating parsing into the rust standard library.

>  If you want a minimal, performant float parser, recent versions of the Rust standard library should be sufficient.

source - [lexical readme](https://github.com/Alexhuszagh/rust-lexical)