@MichaReiser 2022-01-19
I see. Thanks for that explanation. Would you mind running `cargo run -p xtask --release -- coverage-libs` on your PR and on main and doing a short comparison of the numbers? Just to make sure the performance isn't changing in significant ways.