@Boshen 2022-01-19
The benchmark numbers are not stable enough on my laptop, I'll tweak it a little bit and add a CI pipeline first. 

Here are the numbers
<details>

Main
```
jquery.min.js,Total Time,13.32219ms,tokenization,1.640489ms,parsing,4.793683ms,tree_sink,6.888018ms
vue.global.prod.js,Total Time,46.345712ms,tokenization,3.334802ms,parsing,10.961364ms,tree_sink,32.049546ms
react.production.min.js,Total Time,1.85693ms,tokenization,206.719µs,parsing,603.041µs,tree_sink,1.04717ms
react-dom.production.min.js,Total Time,22.263421ms,tokenization,2.867268ms,parsing,8.376348ms,tree_sink,11.019805ms
compiler.js,Total Time,90.405669ms,tokenization,15.06619ms,parsing,25.662806ms,tree_sink,49.676673ms
dojo.js,Total Time,4.192068ms,tokenization,746.663µs,parsing,1.090324ms,tree_sink,2.355081ms
d3.min.js,Total Time,97.490593ms,tokenization,10.42295ms,parsing,36.157603ms,tree_sink,50.91004ms
pixi.min.js,Total Time,65.705268ms,tokenization,7.113466ms,parsing,24.227669ms,tree_sink,34.364133ms
three.min.js,Total Time,235.84244ms,tokenization,19.003818ms,parsing,94.753942ms,tree_sink,122.08468ms
tex-chtml-full.js,Total Time,129.557158ms,tokenization,15.102727ms,parsing,47.942378ms,tree_sink,66.512053ms
```

PR
```
jquery.min.js,Total Time,14.899652ms,tokenization,1.624065ms,parsing,5.225789ms,tree_sink,8.049798ms
vue.global.prod.js,Total Time,22.450082ms,tokenization,2.906751ms,parsing,7.806339ms,tree_sink,11.736992ms
react.production.min.js,Total Time,1.016364ms,tokenization,134.591µs,parsing,361.006µs,tree_sink,520.767µs
react-dom.production.min.js,Total Time,19.957924ms,tokenization,2.231996ms,parsing,6.615241ms,tree_sink,11.110687ms
compiler.js,Total Time,77.698473ms,tokenization,12.395724ms,parsing,20.98178ms,tree_sink,44.320969ms
dojo.js,Total Time,4.136444ms,tokenization,780.5µs,parsing,1.063425ms,tree_sink,2.292519ms
d3.min.js,Total Time,51.583328ms,tokenization,6.316146ms,parsing,18.082167ms,tree_sink,27.185015ms
pixi.min.js,Total Time,67.332459ms,tokenization,7.660642ms,parsing,23.269334ms,tree_sink,36.402483ms
three.min.js,Total Time,74.001429ms,tokenization,8.057979ms,parsing,26.817153ms,tree_sink,39.126297ms
tex-chtml-full.js,Total Time,129.09978ms,tokenization,14.821407ms,parsing,42.808025ms,tree_sink,71.470348ms
```
</details>