@MichaReiser 2022-01-19
> The benchmark numbers are not stable enough on my laptop, I'll tweak it a little bit and add a CI pipeline first.
> 
> Here are the numbers

I normally do a quick comparison of the `criterion` result to see if there are any major regression. The criterion results are more stable because criterion performs multiple runs.