@MichaReiser 2022-01-20
Whoa, this table is amazing. Thanks for the comparison!